refactor: extract NostrConnect handlers into AppState extension (Stage 3a)
Stage 3a of the AppState god-object split — the smallest, lowest-risk
of the three Stage 3 sub-extractions (NostrConnect, ProfileFetcher,
ProxyClient). Establishes the extension-with-mutating-state pattern
for the remaining Stage 3 work.

Moves out of Clave/AppState.swift:
- handleDeeplink(url:) — DeeplinkRouter wiring + dispatch to
  pendingNostrconnectURI vs pendingDeeplinkAccountChoice based on
  account count
- handleNostrConnect(parsedURI:permissions:boundAccountPubkey:) — the
  full multi-relay nostrconnect handshake (~143 LOC): parallel connect,
  publish-with-retry, listen, dispatch RPCs to LightSigner, fire
  pairClientWithProxy on first success
- "// MARK: - Deeplink handling (Task 3.3)" header

Into the new file:
- Clave/AppState+NostrConnect.swift (180 LOC) — extension AppState

What stays in AppState.swift:
- The 3 @Observable stored properties (pendingNostrconnectURI,
  pendingDeeplinkAccountChoice, deeplinkBoundAccount). Swift forbids
  stored properties in extensions. Pointer comment added so readers
  can find the handlers.
- The NotificationCenter observer in AppState.init that calls
  self.handleDeeplink(url:) — works fine as an extension method call.

Zero behavior change:
- Function bodies preserved byte-for-byte (no logic edits, only file
  relocation).
- @MainActor on handleDeeplink preserved.
- All Shared/ dependencies (RelayUtils from Stage 1, LightSigner,
  LightCrypto, LightEvent, SharedKeychain, SharedStorage, etc.) work
  unchanged from the extension.
- pairClientWithProxy(...) call inside handleNostrConnect resolves to
  the extension's `self` (still on AppState until Stage 3c).

External callers unchanged:
- ClaveApp.swift posts .deeplinkReceived to NotificationCenter; the
  observer in AppState.init dispatches to handleDeeplink.
- HomeView.swift and ConnectSheet.swift call appState.handleNostrConnect
  directly — works as an extension method on AppState.

No pbxproj edits — Clave/ directory is a PBXFileSystemSynchronizedRootGroup,
so the new file is picked up automatically by the Clave target.

Verification:
- xcodebuild test -skip-testing:ClaveUITests on iPhone 17 / iOS 26.4:
  - Pre-baseline (main @ 81576e7):  232 passed / 0 failed
  - Post-extraction (this branch):  232 passed / 0 failed
  - ** TEST SUCCEEDED ** in both runs

AppState.swift: 1,934 -> 1,770 LOC (-164, -8.5%).
Combined Stages 1+2+3a: 2,338 -> 1,770 LOC (-24.3% from original).

Stage 3b (ProfileFetcher) and 3c (ProxyClient) follow.
